Get rid of isLazyGitVisible variable
This solves the problem that the variable could run out of sync when hiding the
lazygit window by manually activating a different editor tab; in this case the
extension would still believe that lazygit was active. We could solve this by
implementing onDidChangeActiveTextEditor() to catch this, but why: if we can
determine our visible state without having to remember it, that's better.